+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `ToString` trait implementation for validators.
+- Define `JSONSchema::options` to customise `JSONSchema` compilation [#131](https://github.com/Stranger6667/jsonschema-rs/issues/131)
 
 ### Fixed